requirements.txt
streamlit
pillow
numpy
opencv-python
ibm-watson
